Extract tree expand/collapse logic as hook
